Struct cask::Cask

pub struct Cask { /* fields omitted */ }

An handle to a Cask database.

This handle can be "cheaply" cloned and safely shared between threads. Casks cannot be used concurrently by separate processes and this is ensured by using a file lock in the Cask dir.

Methods

impl Cask

pub fn open(path: &str, options: CaskOptions) -> Result<Cask>

Opens/creates a new Cask.

pub fn compact(&self) -> Result<()>

Trigger Cask log compaction.

pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<Vec<u8>>>

Returns the value corresponding to the key, if any.

pub fn put<K: Into<Vec<u8>>, V: AsRef<[u8]>>(
&self,
key: K,
value: V
) -> Result<()>

Inserts a key-value pair into the map.

pub fn delete<K: AsRef<[u8]>>(&self, key: K) -> Result<()>

Removes a key from the map.